1.2.0 (?)

+++ PVPPremium/src/com/Sagacious_/KitpvpStats/bounty/BountyManager.java
@@ -4,6 +4,8 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -35,7 +37,7 @@ public class BountyManager {
 	private File f;
 	private File dir;
 	public FileConfiguration conf;
-	private int time;
+	public int time;
 	private Sound bounty;
 	public Sound survived;
 	public Sound kill;
@@ -83,6 +85,9 @@ public class BountyManager {
 			bg = new BountyGUI();
 			Bukkit.getPluginManager().registerEvents(bg, Core.getInstance());
 			new CommandBounty();
+			if(Core.getInstance().sql!=null) {
+				updateBounties();
+			}
 			}
 		}
 	}
@@ -104,7 +109,41 @@ public class BountyManager {
 		return s;
 	}
 	
+	private void updateBounties() {
+		for(Bounty b : getBounties()) {
+			try {
+				ResultSet set = Core.getInstance().sql.prepareStatement("SELECT * FROM Bounty WHERE id='" + b.id + "'").executeQuery();
+				Bounty s = new Bounty(set.getString("id"), UUID.fromString(set.getString("uuidOwner")), UUID.fromString(set.getString("uuidTarget")), set.getDouble("amount"), set.getInt("timeLeft"));
+				set.close();
+				if(s.getTimeLeft()>b.getTimeLeft()) {
+			    b.save(false);
+				}else {
+				replaceBounty(b, s);
+				}
+			} catch (SQLException e) {
+				e.printStackTrace();
+			}
+		}
+	}
+	
+	private void replaceBounty(Bounty first, Bounty second) {
+		UUID z = null;
+		for(Entry<UUID, List<Bounty>> s : bounties.entrySet()) {
+			for(Bounty f : s.getValue()) {
+				if(f.equals(first)) {
+					z=s.getKey();
+				}
+			}
+		}
+		if(z!=null) {
+			List<Bounty> q = bounties.get(z);
+			q.remove(first);q.add(second);
+			bounties.put(z, q);
+		}
+	}
+	
 	private void loadAll() {
+		if(Core.getInstance().sql==null) {
 		FileConfiguration confe;
 		for(File z : dir.listFiles()) {
 			confe = YamlConfiguration.loadConfiguration(z);
@@ -119,11 +158,26 @@ public class BountyManager {
 				bounties.put(b.getUuidTarget(), new ArrayList<Bounty>(Arrays.asList(b)));
 			}
 		}
+		}else {
+			try {
+				ResultSet set = Core.getInstance().sql.prepareStatement("SELECT * FROM Bounty").executeQuery();
+				while(set.next()) {
+					Bounty b = new Bounty(set.getString("id"), UUID.fromString(set.getString("uuidOwner")), UUID.fromString(set.getString("uuidTarget")), set.getDouble("amount"), set.getInt("timeLeft"));
+				    createBounty(b.getUuidOwner(), b.getUuidTarget(), b.getAmount(), b);
+				}
+				set.close();
+			} catch (SQLException e) {
+				e.printStackTrace();
+			}
+		}
 	}
 	
-	public void createBounty(UUID owner, UUID target, double amount) {
+	public void createBounty(UUID owner, UUID target, double amount, Bounty z) {
 		String id = generateID();
-		Bounty b = new Bounty(id, owner, target, amount, time);
+		Bounty b = null;
+		if(z==null) {b=new Bounty(id, owner, target, amount, time);}else {
+			b=z;
+		}
 		if(bounties.containsKey(b.getUuidTarget())) {
 			List<Bounty> s = bounties.get(b.getUuidTarget());
 			s.add(b);
@@ -131,7 +185,9 @@ public class BountyManager {
 		}else {
 			bounties.put(b.getUuidTarget(), new ArrayList<Bounty>(Arrays.asList(b)));
 		}
+		if(z==null) {
 		playSound(Bukkit.getPlayer(target), bounty);
+		}
 	}
 	
 	public List<Bounty> getBounties(){
@@ -146,7 +202,7 @@ public class BountyManager {
 		if(!bounties.isEmpty()) {
 		for(Entry<UUID, List<Bounty>> b : bounties.entrySet()) {
 			for(Bounty bo : b.getValue()) {
-				bo.save();
+				bo.save(true);
 			}
 		}
 		}
@@ -159,8 +215,16 @@ public class BountyManager {
 					s.getValue().remove(b);
 				}
 				b.active=false;
+				if(Core.getInstance().sql==null) {
 				File f = new File(dir, b.id + ".yml");
 				f.delete();
+				}else {
+					Bukkit.getScheduler().runTaskAsynchronously(Core.getInstance(), new Runnable() {
+						public void run() {
+							Core.getInstance().sql.executeStatement("DELETE * FROM Bounty WHERE id='" + b.id + "';");
+						}
+					});
+				}
 			}
 		}
 	}
